+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for converting singular blp file into png
 - Auto-detect correct blp compression for world map textures
+- Automation scripts
 
 ### Fixed